requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-jose[cryptography]
pydantic
passlib[bcrypt]
sqlalchemy